split the filter engine tests
